+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [1.2.14] = 8 August 2022
 ### Fixed
 - Properly pick up on injected interfaces too
+- Support multiple namespace tokens
+- Fix token warning on PHP 7.4
 
 # [1.2.13] = 8 August 2022
 ### Removed